+++ examples/PerfTest1/PerfTest1.ino
@@ -6,8 +6,8 @@
 #endif
 #if SOC_UART_HP_NUM < 3
   #define Serial2 Serial1
-  #define RX2 RX1
-  #define TX2 TX1
+  #define RX2     RX1
+  #define TX2     TX1
 #endif
 
 // Pin: RX (ESP32) -> TX (JSY)
@@ -55,13 +55,13 @@ void setup() {
     jsy.end();
     jsy.begin(Serial2, RX2, TX2);
 
-    Serial.printf("\njsy.read() at %d bauds:\n", static_cast<int>(rates[i]));
+    Serial.printf("\njsy.read() at %" PRIu32 " bauds:\n", rates[i]);
 
     // Serial.printf(" -");
     int64_t times[50];
     for (size_t j = 0; j < 50; j++) {
       times[j] = esp_timer_get_time();
-      if(jsy.read()) {
+      if (jsy.read()) {
         times[j] = esp_timer_get_time() - times[j];
       } else {
         times[j] = 0;
